changed upkeep

+++ Yakuza_Syndicate/Yakuza_Syndicate/GangMembers.h
@@ -7,7 +7,7 @@ private:
 	sf::Font font;
 
 	int amount;
-	int upkeep;
+	float upkeep;
 	bool action;
 	bool inFriendlyTerr;
